:art: extract MeterArrow
